+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next
 
 - Add "Return to top" component
+- Add `dollarFormat` utility function
+- Add `kebabCase` utility function
 
 ## 0.2.1